created a new ApiClientUtils folder and moved everything to that folder
